Added to file system method for check on exist file or directory
git-svn-id: svn://192.168.0.101/eleot@118 7f037352-864a-4b62-bc5a-4f07d5e00ced

+++ Develompent/Source/Engine/WorldEd/Source/Commandlets/CookPackagesCommandlet.cpp
@@ -656,7 +656,10 @@ bool LCookPackagesCommandlet::Main( const std::wstring& InCommand )
 
 	// Clear table of content and if cooked dir already created remove it
 	GTableOfContents.Clear();
-	GFileSystem->DeleteDirectory( GCookedDir, true );
+	if ( GFileSystem->IsExistFile( GCookedDir, true ) )
+	{
+		GFileSystem->DeleteDirectory( GCookedDir, true );
+	}
 
 	// Cook all resource with flag bAlwaysCook = true
 	CookAllResources( true );
